Refactored Navbar component into functional component. Split NavList into its own component.
