+++ src/components/Navbar.js
@@ -1,27 +1,24 @@
-import React, { Component } from 'react'
+import React, { memo, useState, useEffect } from 'react'
 import PropTypes from 'prop-types'
-import styled, { css } from 'styled-components'
+import NavList from './NavList'
+import styled from 'styled-components'
 import { theme, Keyframes as KF, SquareOutlineIcon } from '@theme'
-import { generateKey } from '@utils'
 import debounce from 'lodash.debounce'
-import { navLinks } from '@data/site.yml'
 
-class Navbar extends Component {
-    hiddenNavListItems = ['Home']
-    debounceInteral = 50
+// =============================================================================
 
-    state = {
+const Navbar = ({ mainContentRef }) => {
+    const [state, setState] = useState({
         isOpen: false,
         isHidden: false,
         scrollPosition: 0,
-    }
+    })
 
-    toggleNavList = () => this.setState({ isOpen: !this.state.isOpen })
+    const toggleNavList = () => setState({ ...state, isOpen: !state.isOpen })
+    const closeNavList = () => setState({ ...state, isOpen: false })
 
-    closeNavList = () => this.setState({ isOpen: false })
-
-    handleScroll = debounce(() => {
-        this.setState((prevState) => {
+    const handleScroll = debounce(() => {
+        setState((prevState) => {
             const scrollPosition = document.body.getBoundingClientRect().top
 
             return {
@@ -31,64 +28,35 @@ class Navbar extends Component {
                 isOpen: false,
             }
         })
-    }, this.debounceInteral)
+    }, debounceInteral)
 
-    componentDidMount() {
-        window.addEventListener('scroll', this.handleScroll)
-    }
-
-    componentDidUpdate(prevProps) {
-        const { mainContentRef } = prevProps
+    useEffect(() => {
+        window.addEventListener('scroll', handleScroll)
 
         if (mainContentRef.current) {
-            mainContentRef.current.addEventListener('click', this.closeNavList)
+            mainContentRef.current.addEventListener('click', closeNavList)
         }
-    }
 
-    componentWillUnmount() {
-        window.removeEventListener('scroll', this.handleScroll)
-    }
+        return () => {
+            window.removeEventListener('scroll', handleScroll)
+            mainContentRef.current.removeEventListener('click', closeNavList)
+        }
+    }, [])
 
-    render() {
-        const { state, hiddenNavListItems, toggleNavList, closeNavList } = this
-        const { isOpen, isHidden } = state
-
-        return (
-            <Header className={isHidden && 'hidden'}>
-                <Nav>
-                    <MenuButtonContainer onClick={toggleNavList}>
-                        <SquareOutlineIcon />
-                    </MenuButtonContainer>
-                    <Logo href="#">Kevin Han</Logo>
-                    <NavList isOpen={isOpen}>
-                        {navLinks.map((link, i) => {
-                            const { text, ...linkProps } = link
-                            const navListItemProps = {}
-
-                            if (hiddenNavListItems.includes(text)) {
-                                navListItemProps.additionalStyles = additionalHiddenNavListItemStyles
-                            }
-
-                            return (
-                                <NavListItem
-                                    key={generateKey(text, i)}
-                                    {...navListItemProps}
-                                >
-                                    <NavLink
-                                        {...linkProps}
-                                        onClick={closeNavList}
-                                    >
-                                        {text}
-                                    </NavLink>
-                                </NavListItem>
-                            )
-                        })}
-                    </NavList>
-                </Nav>
-            </Header>
-        )
-    }
+    return (
+        <Header className={state.isHidden && 'hidden'}>
+            <Nav>
+                <MenuButtonContainer onClick={toggleNavList}>
+                    <SquareOutlineIcon />
+                </MenuButtonContainer>
+                <Logo href="#">Kevin Han</Logo>
+                <NavList isOpen={state.isOpen} closeNavList={closeNavList} />
+            </Nav>
+        </Header>
+    )
 }
+
+const debounceInteral = 50
 
 Navbar.propTypes = {
     mainContentRef: PropTypes.object,
@@ -153,6 +121,8 @@ const Logo = styled('a')`
 // =============================================================================
 
 const MenuButtonContainer = styled('button')`
+    background: transparent;
+    border: 0;
     cursor: pointer;
 
     svg {
@@ -189,121 +159,10 @@ const Nav = styled('nav')`
 
 // =============================================================================
 
-const NavList = styled('ul')`
-    position: absolute;
-    top: 100%;
-    left: 0;
-    display: ${({ isOpen }) => (isOpen ? 'block' : 'none')};
-    width: 100%;
-    padding: 0 16px;
-    list-style: none;
-    background: rgba(${theme.colorBackground}, 1);
+export default memo(Navbar, (prevProps, nextProps) => {
+    const isSameMainContentRefProp =
+        prevProps.mainContentRef.current &&
+        prevProps.mainContentRef.current === nextProps.mainContentRef.current
 
-    @media only screen and (min-width: 768px) {
-        position: relative;
-        top: unset;
-        left: unset;
-        display: flex;
-        width: auto;
-        height: 100%;
-        padding: 0;
-        /* background: rgba(${theme.colorBackground}, 0.98); */
-        border: 0;
-    }
-`
-
-NavList.defaultProps = {
-    isOpen: false,
-}
-
-NavList.propTypes = {
-    isOpen: PropTypes.bool,
-}
-
-// =============================================================================
-
-const NavListItem = styled('li')`
-    @media only screen and (min-width: 768px) {
-        opacity: 0;
-
-        & + & {
-            margin-left: 46px;
-        }
-
-        &:nth-child(2) {
-            animation: ${KF.slide} 500ms linear 1750ms 1 forwards;
-        }
-
-        &:nth-child(3) {
-            animation: ${KF.slide} 500ms linear 1900ms 1 forwards;
-        }
-
-        &:nth-child(4) {
-            animation: ${KF.slide} 500ms linear 2050ms 1 forwards;
-        }
-    }
-
-    @media only screen and (min-width: 1024px) {
-        & + & {
-            margin-left: 50px;
-        }
-    }
-
-    @media only screen and (min-width: 1024px) {
-        & + & {
-            margin-left: 60px;
-        }
-    }
-
-    ${(props) => props.additionalStyles}
-`
-
-const additionalHiddenNavListItemStyles = css`
-    @media only screen and (min-width: 768px) {
-        display: none;
-    }
-`
-
-// =============================================================================
-
-const NavLink = styled('a')`
-    display: flex;
-    padding: 16px 0;
-    font-size: 16px;
-
-    @media only screen and (min-width: 768px) {
-        position: relative;
-        align-items: center;
-        height: 100%;
-        padding: 0;
-        transition: color ${theme.transitionNavLink};
-
-        /* Defines the underline at the bottom of each navigation list item. */
-        &::after {
-            position: absolute;
-            bottom: 0;
-            left: 0;
-            width: 100%;
-            height: 3px;
-            background: rgba(${theme.colorPrimary}, 1);
-            transform: scaleX(0);
-            transform-origin: right;
-            transition: transform ${theme.transitionNavLink};
-            content: '';
-        }
-
-        &:hover,
-        &:active {
-            color: rgba(${theme.colorHeading}, 0.6);
-
-            &::after {
-                transform: scaleX(1);
-                transform-origin: left;
-            }
-        }
-    }
-`
-
-// =============================================================================
-
-export default Navbar
+    return isSameMainContentRefProp
+})
